refactor(crouton-layout): move the layout-engine implementation out of core (#754, #755)
👤 For humans
Moves the layout engine's code — the editor/renderer components, the
composables, the pure layout utils, and the default blocks — out of
crouton-core and into the @fyit/crouton-layout layer. Behaviour is
unchanged: the components keep their exact Crouton* names and are auto-on
everywhere (crouton-layout is already default-on, #757), so no app needed a
change. The server side + the harness page follow in #756.

🤖 For agents
Combines #754 (utils) and #755 (components+composables): they import each
other by relative path, so splitting them would need a backward core→layout
shim (ruled out by the greenfield/no-compat decision, epic #751).

Moved → packages/crouton-layout:
- app/utils: layout-compose, layout-viability, layout-edit (+ __tests__).
- app/composables: useCroutonLayout{Blocks,Edit,Store} (+ the Blocks test).
- app/components: Layout, LayoutRenderer, LayoutEditorPane, LayoutCollection
  (+Data), LayoutForm, LayoutSpikeStats (Crouton* names preserved via the
  layer's components.dirs prefix).
- app/app.config.ts: the croutonLayoutBlocks defaults (collection-list /
  entity-form / stats), defu-merged across layers.
- vitest.config.ts + test script (root vitest auto-discovers packages/*/vitest.config.ts).

Stayed in crouton-core (shared contract — so bookings/pages need NO dep on
crouton-layout):
- app/types/layout.ts, app/types/layout-block.ts, app/types/block-core.ts.
- app/utils/layout-tree.ts (sanitizeLayoutTree) — core's server PUT still uses
  it; moves to crouton-layout with the server in #756.
- app/composables/useCroutonBlockCatalog.ts (bridges both block registries).

Wiring:
- Moved files import the layout TYPES from @fyit/crouton-core/app/types/...;
  useCroutonLayoutStore imports sanitizeLayoutTree from core. One-way deps
  (crouton-layout → crouton-core) preserved; no backward edges.
- CLI lib/compose-layout.ts repointed to @fyit/crouton-layout/app/utils/layout-compose
  (the type import stays on core).
- core package.json: dropped the moved ./app/utils/layout-{compose,viability}
  exports; kept the type + layout-tree exports.

Verified: crouton-layout 57 tests + core 690 tests pass; CLI 267 tests pass
(compose repoint works in jiti); per-app typecheck green (24 manifests);
e2e-fixture minimal + with-bookings typecheck exit 0 (cross-package
bookings-calendar block still resolves). Core CLAUDE.md row cleanup + the
full boot/e2e proof are #758's scope.
