+++ packages/crouton-layout/app/components/LayoutCollectionData.vue
