+++ packages/crouton-layout/app/components/LayoutSpikeStats.vue
