+++ packages/crouton-layout/package.json
@@ -5,7 +5,14 @@
   "type": "module",
   "main": "./nuxt.config.ts",
   "exports": {
-    ".": "./nuxt.config.ts"
+    ".": "./nuxt.config.ts",
+    "./app/utils/layout-compose": "./app/utils/layout-compose.ts",
+    "./app/utils/layout-viability": "./app/utils/layout-viability.ts",
+    "./app/utils/layout-edit": "./app/utils/layout-edit.ts"
+  },
+  "scripts": {
+    "test": "vitest run",
+    "test:watch": "vitest"
   },
   "files": [
     "app",
@@ -39,5 +46,10 @@
     "@fyit/crouton-core": "workspace:*",
     "@nuxt/ui": "^4.3.0",
     "nuxt": "^4.0.0"
+  },
+  "devDependencies": {
+    "@vitest/coverage-v8": "^2.1.9",
+    "happy-dom": "^16.6.0",
+    "vitest": "^2.1.0"
   }
 }
